Not all tiles used case
Algorithm used now has a decision tree based on input. If the biggest
tile is over half the puzzle, old algorithm is used, if not all tiles
are used, subset algorithm is used, otherwise, new algorithm is used.
Parser.rotate_solution indicates whether or not the solutions need to be
rotated clockwise 90 degrees for display